+++ examples/tutorials/lighting/app-old.ts
@@ -1,6 +1,12 @@
 import {glsl, NumberArray, UniformStore, ShaderUniformType} from '@luma.gl/core';
 import type {AnimationProps} from '@luma.gl/engine';
-import {AnimationLoopTemplate, Model, CubeGeometry, loadImage, AsyncTexture} from '@luma.gl/engine';
+import {
+  AnimationLoopTemplate,
+  Model,
+  CubeGeometry,
+  loadImage,
+  DynamicTexture
+} from '@luma.gl/engine';
 import {phongMaterial, lighting} from '@luma.gl/shadertools';
 import {Matrix4} from '@math.gl/core';
 
@@ -125,7 +131,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
       }
     });
 
-    const texture = new AsyncTexture(device, {data: loadImage('vis-logo.png')});
+    const texture = new DynamicTexture(device, {data: loadImage('vis-logo.png')});
 
     this.model = new Model(device, {
       vs,